refactor: create subcomponents in ProfileDeposit.tsx for re-use in oneTimeReward
